Safari-609.1.11.1/LayoutTests/inspector/unit-tests/promise-utilities-expected.txt


== Running test suite: Promise
-- Running test case: Promise.chain
PASS: Resolved promise 0 with value 1
PASS: Resolved promise 1 with value 2
PASS: Resolved promise 2 with value 4
